requirements.txt
ruamel.yaml
networkx
